--- anycompany/src/main/java/za.co.anycompany/model/Order.java
@@ -1,32 +1,0 @@
-package za.co.anycompany.model;
-
-public class Order {
-
-    private int orderId;
-    private double amount;
-    private double VAT;
-
-    public int getOrderId() {
-        return orderId;
-    }
-
-    public void setOrderId(int orderId) {
-        this.orderId = orderId;
-    }
-
-    public double getAmount() {
-        return amount;
-    }
-
-    public void setAmount(double amount) {
-        this.amount = amount;
-    }
-
-    public double getVAT() {
-        return VAT;
-    }
-
-    public void setVAT(double VAT) {
-        this.VAT = VAT;
-    }
-}
